+++ src/parser/monster/source.js
@@ -2,15 +2,17 @@ import DDBHelper from "../../lib/DDBHelper.js";
 import DDBMonster from "../DDBMonster.js";
 
 DDBMonster.prototype._generateSource = function _generateSource() {
-  const sourceObject = CONFIG.DDB.sources.find((cnf) => cnf.id == this.source.sourceId);
-  const book = (sourceObject)
-    ? sourceObject.name
-    : "Homebrew";
+  const ddbSource = CONFIG.DDB.sources.find((cnf) => cnf.id == this.source.sourceId);
 
-  this.npc.system.details.source = DDBHelper._tweakSourceData({
-    book,
-    page: this.source.sourcePageNumber,
+  const source = {
+    book: ddbSource ? ddbSource.name : "Homebrew",
+    page: this.source.sourcePageNumber ?? "",
     custom: "",
     license: "",
-  });
+    id: ddbSource ? ddbSource.id : 9999999,
+  };
+
+  DDBHelper._tweakSourceData(source);
+
+  this.npc.system.details.source = source;
 };