+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix for weapons such as Frost Band weapons not  generating extra elemental damage.
 - Healing Spells would sometimes have a utility activity when it was not needed.
 - Spell Tasha's Bubbling Cauldron spell tweaks.
+- Monster book sources had stopped importing.
 
 # 6.0.8